requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
requests
beautifulsoup4
